Initial commit
Includes library for simulation of diffusion types, a library to compute diffusional fingerprints and a library to train machine learning classifiers and analyze results. Finally, a script called "Usage_example.py" is included to show how the various functions may be used to perform analysis like the ones presented in the paper.